Add error handling to burndown chart
As I touched this, I figured I'd also do the migration to apollo.
